+++ internal/internalapi/internalapi.go
@@ -99,35 +99,19 @@ const (
 
 	// FilterOrderAnnotation is the annotation key on an EnvoyExtensionPolicy that controls
 	// the placement of its extension filters relative to the AI Gateway ext_proc filter.
-	//
 	// The value is a comma-separated ordered sequence of filter type tokens with the special
 	// token "ExtProc" acting as the pivot that represents the AI Gateway ext_proc. Filters
 	// listed before "ExtProc" are placed before it in the HCM chain; filters listed after
 	// "ExtProc" are placed after it. Filters not mentioned keep their current relative order.
-	//
-	// Recognised tokens (case-insensitive):
-	//   Lua, Wasm, DynamicModules, ExtProc, rbac, LocalRateLimit, RateLimit
-	//
-	// Example: "Wasm,Lua,ExtProc,LocalRateLimit"
-	//   → Wasm and Lua run before the AI Gateway ext_proc on requests
-	//     (after it on responses, because Envoy reverses response direction).
-	//   → LocalRateLimit runs after the AI Gateway ext_proc on requests.
-	//
-	// Backward-compatible value: "before-extproc" — moves all of this policy's extension
-	// filters before the AI Gateway ext_proc (equivalent to "Lua,Wasm,DynamicModules,ExtProc").
 	FilterOrderAnnotation = "aigateway.envoyproxy.io/filter-order"
 
-	// FilterOrderBeforeExtProc is a legacy value for FilterOrderAnnotation that places all
-	// of the policy's extension filters before the AI Gateway ext_proc. Retained for
-	// backward compatibility; prefer the sequence format for new configs.
+	// FilterOrderBeforeExtProc is the value for DefaultFilterOrderAnnotation that places all
+	// of the policy's extension filters before the ext_proc.
 	FilterOrderBeforeExtProc = "before-extproc"
 
 	// DefaultFilterOrderAnnotation is the default annotation key for controlling filter order,
-	// targeting Lua filters specifically. New configs should use FilterOrderAnnotation instead.
+	// targeting Lua filters specifically.
 	DefaultFilterOrderAnnotation = "aigateway.envoyproxy.io/lua-filter-order"
-
-	// DefaultFilterOrderBeforeExtProc is the default value for DefaultFilterOrderAnnotation.
-	DefaultFilterOrderBeforeExtProc = FilterOrderBeforeExtProc
 )
 
 // ParseRequestHeaderAttributeMapping parses comma-separated key-value pairs for header-to-attribute mapping.
