+++ tests/e2e/testdata/lua_filter_order.yaml
@@ -157,7 +157,7 @@ metadata:
   name: lua-filter-order
   namespace: default
   annotations:
-    aigateway.envoyproxy.io/lua-filter-order: post-extproc
+    aigateway.envoyproxy.io/filter-order: Lua,ExtProc
 spec:
   targetRefs:
     - group: gateway.networking.k8s.io
